Changed price field type.

+++ src/SecondWeek/homework1/entity/House.java
@@ -2,7 +2,7 @@ package SecondWeek.homework1.entity;
 
 public class House extends Housing {
 
-    public House(Double price, int squareMeter, int numberOfRooms, int numberOfLivingRooms) {
+    public House(Long price, int squareMeter, int numberOfRooms, int numberOfLivingRooms) {
         super(price, squareMeter, numberOfRooms, numberOfLivingRooms);
     }
